Proper authz for COPY and MOVE.
WebDAV copy and move need to be authorized in two context, the source
file context and the destination one (provided via the destination
header).

The current authz approach supports cross storage area COPY and MOVE
operations.

Authz on the source is done using spring provided request matchers.
Authz on the destination is implemented in CopyMoveAuthzVoter, which
looks into the Destination header and checks whether the requestor has
proper permissions in the target storage area.