Tentative fix for STOR-738
Previously storm-webdav raised an exception that resulted in a 500 http
status when trying to move a file over itself. While harmless in
practice, a 403 should be returned in this case, according to the webdav
spec.

The problem is that Milton does not check for identical source and
destination URIs in a MOVE, so the check must be done in the underlying
code. Moreover, Milton proactively removes the destination URI before
doing the move if the user is authorized and the Overwrite header is set
to T, so what happens is that a move of a file over itself results in a
delete in practice.

Fortunately, the removal of  the destination resource before doing the
move can be disabled setting a flag in the the Milton's MoveHandler
code. This flag cannot be easily set through the builder, so we plug in
(in an hackish way) right after the webdav protocol initialization.