Support for OIDC/OAuth token issuers
The init.d script is modified to change directory to /etc/storm/webdav,
so that we can leverage spring-boot support for overriding the embedded default
configuration. This is convenient to provide the list of trusted
OAuth/OIDC token issuers in an intuitive way.

Added preliminary doc via a README.md file.