Redirector support
Added redirector support, i.e.,  the ability to redirect GET and PUT
requests to other transfer nodes, including an access token in the
redirect URL to grant file-level, time-limited authorization to the
bearer of the token.

This functionality allows, for instance, to handle data transfer in
plain HTTP to reduce load on transfer servers or to delegate the
transfer to a pool of nodes, using either the http or https protocol.

This commit also introduces:

- traceability information in the access log
- storage area ok/error count request metrics